Fixes #7 : Other is added as automatic option
For consistency, menu option "Other" is added to the Assay Type menu. If selected, user is asked to enter a new Assay Type ID. A new AssayType directory is created as a skeleton in templates. This enables easy extensions to new types. This is valid for dry and wet assays.