major redesign with x.lib
This is a major redesign of a system.

all batch content is packed in lib.cmd file. Make* batch files are calling the appropriate rutines from the lib.cmd file. They are generic in a way, that no future updates will be neccessary, as long as the locaton of the lib.cmd file will remain fixed.